changed layer height to float to support decimal layer heights, added initial non-printed layer of 3 mm fixed, changed depositing to two times half a Z distance to prevent bow-wave

+++ main.py
@@ -19,7 +19,7 @@ print(f"Selected file: {file_path}")
 print()
 frequency = int(input("Enter valve frequency [Hz]: "))
 duty_cycle = float(input("Enter valve duty cycle [0-1]: "))
-layer_height = int(input("Enter layer height [mm]: "))
+layer_height = float(input("Enter layer height [mm]: "))
 
 h.process(frequency, duty_cycle, layer_height)
